refactor: separate baby meal detail into smaller files
